requirements.txt
beautifulsoup4==4.12.2
docx2txt==0.8
httpx==0.25.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
nltk==3.8.1
numpy==1.26.2
openai==0.28.0
pandas==2.1.3
pdfminer==20191125
Pillow>=10.3.0
pipenv==2023.11.15
plotly==5.18.0
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
PyPDF2>=2.10.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.0
streamlit>=1.30.0
streamlit-authenticator==0.3.1
stripe==8.0.0
textstat==0.7.3
toml==0.10.2
